Output variables for SIBCASA

TODO: Provide a complete description here of the dataset, including its origin and purpose. The column descriptions below were written by someone looking at the file with no insight into its usage

sibcasa_output_vars

Format

data frame with 6 unnamed columns

1

? (t 0, t 3, etc)

2

short name

3

row number?

4

description

5

units

6

unused